+++ back-end/NoteApi/Service/INoteService.cs
@@ -5,15 +5,15 @@ namespace NoteApi.Service
     public interface INoteService
     {
         Task<IEnumerable<CommonNoteVM>> GetCommonNotesAsync();
-        Task<IEnumerable<CommonNoteVM>> GetCommonNotesAsync(int userId);
+        Task<IEnumerable<CommonNoteVM>> GetCommonNotesAsync(long userId);
         Task<IEnumerable<PlanNoteVM>> GetPlanNotesAsync();
-        Task<IEnumerable<PlanNoteVM>> GetPlanNotesAsync(int userId);
+        Task<IEnumerable<PlanNoteVM>> GetPlanNotesAsync(long userId);
         Task<IEnumerable<SpecialityNoteVM>> GetSpecialityNotesAsync();
-        Task<IEnumerable<SpecialityNoteVM>> GetSpecialityNotesAsync(int userId);
+        Task<IEnumerable<SpecialityNoteVM>> GetSpecialityNotesAsync(long userId);
         Task<IEnumerable<SpecialityHighlightNoteVM>> GetSpecialityHighlightNotesAsync();
-        Task<IEnumerable<SpecialityHighlightNoteVM>> GetSpecialityHighlightNotesAsync(int userId);
+        Task<IEnumerable<SpecialityHighlightNoteVM>> GetSpecialityHighlightNotesAsync(long userId);
         Task<IEnumerable<SpecialityPriorityNoteVM>> GetSpecialityPriorityNotesAsync();
-        Task<IEnumerable<SpecialityPriorityNoteVM>> GetSpecialityPriorityNotesAsync(int userId);
+        Task<IEnumerable<SpecialityPriorityNoteVM>> GetSpecialityPriorityNotesAsync(long userId);
         Task<int> AddCommonNoteAsync(CommonNoteVM note);
         Task<int> AddPlanNoteAsync(PlanNoteVM note);
         Task<int> AddSpecialityNoteAsync(SpecialityNoteVM note);
@@ -28,6 +28,6 @@ namespace NoteApi.Service
         Task DeleteSpecialityNoteAsync(int noteId);
         Task DeleteSpecialityHighlightNoteAsync(int noteId, bool positive, int specialityId);
         Task DeleteSpecialityPriorityNoteAsync(int noteId, int specialityId);
-        Task<IEnumerable<SpecialityRatingVM>> GetSpecialityRatingNoteAsync(int userId);
+        Task<IEnumerable<SpecialityRatingVM>> GetSpecialityRatingNoteAsync(long userId);
     }
 }
